Do not attempt to grab the database twice

+++ ilios-entrypoint.sh
@@ -17,15 +17,17 @@ for arg; do
 	esac
 done
 
-echo 'Retrieving Ilios Demo Database...'
-/usr/bin/wget --no-verbose https://ilios-demo.ucsf.edu/latest_db/ilios3_demosite_db.sql.gz
-echo 'done... unpacking demo database'
-gunzip ilios3_demosite_db.sql.gz
-echo 'done.... copying ilios demo database to by read automatically by docker'
-echo "USE ilios;" > /docker-entrypoint-initdb.d/ilios.sql
-cat ilios3_demosite_db.sql >> /docker-entrypoint-initdb.d/ilios.sql
-rm ilios3_demosite_db.sql
-echo 'done'
+if [ ! -f /docker-entrypoint-initdb.d/ilios.sql ]; then
+	echo 'Retrieving Ilios Demo Database...'
+	/usr/bin/wget --no-verbose https://ilios-demo.ucsf.edu/latest_db/ilios3_demosite_db.sql.gz
+	echo 'done... unpacking demo database'
+	gunzip ilios3_demosite_db.sql.gz
+	echo 'done.... copying ilios demo database to by read automatically by docker'
+	echo "USE ilios;" > /docker-entrypoint-initdb.d/ilios.sql
+	cat ilios3_demosite_db.sql > /docker-entrypoint-initdb.d/ilios.sql
+	rm ilios3_demosite_db.sql
+	echo 'done'
+fi
 
 _datadir() {
 	"$@" --verbose --help 2>/dev/null | awk '$1 == "datadir" { print $2; exit }'
